GenError and GenFile
