+++ presentation-compiler/src/test/scala/org/alephium/ralph/lsp/access/compiler/message/error/GenError.scala
@@ -1,0 +1,24 @@
+package org.alephium.ralph.lsp.access.compiler.message.error
+
+import org.alephium.ralph.lsp.GenCode
+import org.alephium.ralph.lsp.access.compiler.message.{CompilerMessage, SourceIndex}
+import org.scalacheck.Gen
+
+object GenError {
+
+  /** Generate an error for this code */
+  def genError(code: Gen[String] = GenCode.genGoodCode()): Gen[CompilerMessage.AnyError] =
+    for {
+      code <- code
+      errorMessage <- Gen.alphaStr
+      errorIndex <- Gen.choose(0, code.length - 1)
+    } yield
+      StringError(
+        message = errorMessage,
+        index = SourceIndex(0, errorIndex) // TODO: gen random index location
+      )
+
+  def genErrors(code: String): Gen[List[CompilerMessage.AnyError]] =
+    Gen.listOf(genError(Gen.const(code)))
+
+}
